backend/python/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
pdfplumber
pytesseract
Pillow
python-docx
python-pptx
beautifulsoup4
requests
youtube-transcript-api
chromadb
sentence-transformers
google-generativeai
langchain
langchain-community
python-dotenv
langchain-ollama
